requirements.txt
# Core dependencies
streamlit
pandas
openai
pinecone
pymongo==4.6.2
dnspython==2.6.1
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.1
pydantic

# Crawler dependencies
crawl4ai
html2text
lxml

# Utility dependencies
tqdm
aiohttp